rename rpc-state.h -> rpc-client-state.h
